pyproject.toml
[project]
name = "modera-dental-agent"
version = "0.1.0"
description = "LiveKit Voice AI Agent for Modera Dental"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "httpx>=0.28.1",
    "livekit-agents[silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
]
